Remake of Wolfenstein3D, in plain C
This was a side project that was supposed to be a 3 day challenge to
recreate Wolf3D. Instead, we got about 90% done in the 3 days, and then
Coop and I took another 2 weeks to iron out some details and make it
feel as close to the original as possible. Still missing a few features,
but good enough.