Uploaded primary GDK TFS repository.
Contains various incarnations of Lucid3D, GameDevelopmentKit, and other
side projects from the past 2-4 years. This was done together with
Cooper Partin. All copyrighted content has been stripped out to the best
of my knowledge, please let me know if you find anything that should be
removed. You can usually look in the code to see what is being attempted
to be loaded and if you own the game/content place the content at that
location (or fix the path) to run the code (if it even compiles!).